Introduce a `NotificationHandler` object to split out notification content handling from the rest of the NSE core logic.
